app/requirements.txt
flask==3.1.*
gunicorn==23.*
python-dotenv==1.*
requests==2.*
